Move setup_test_module() to test_utils.rs so that it can be used from other crates as well.
