+++ crates/semantic/Cargo.toml
@@ -3,6 +3,9 @@ name = "semantic"
 version = "0.1.0"
 edition = "2021"
 
+[features]
+testing = []
+
 [dependencies]
 db_utils = { path = "../db_utils" }
 defs = { path = "../defs" }
